Send the success rate and token total the panel has been asking for
A project's detail page showed 0% success next to "1 failed" out of 1.6K
requests, and 0 tokens directly above "1.9M in, 96K out". Neither figure was
wrong so much as absent: UsageSummaryDto never carried success_rate or
total_tokens, so the panel read undefined and rendered a zero. The in/out
counts looked right because those two fields did exist.

Both are now computed on the DTO rather than in the panel, since every caller
wants them and none of them should be dividing counts itself. The new project
cards read the same field, so they were about to report every project at 0%.

The token sums widen from int to long while I am here. They are lifetime
totals for a project, and LINQ's Sum throws on overflow rather than wrapping,
so on a busy gateway this would eventually turn into a 500 rather than a wrong
number.

+++ src/Freeway.Application/Features/Analytics/Queries/GetProjectUsageQueryHandler.cs
@@ -40,8 +40,8 @@ public class GetProjectUsageQueryHandler : IRequestHandler<GetProjectUsageQuery,
             TotalRequests = logs.Count,
             SuccessfulRequests = logs.Count(l => l.Success),
             FailedRequests = logs.Count(l => !l.Success),
-            TotalInputTokens = logs.Sum(l => l.InputTokens),
-            TotalOutputTokens = logs.Sum(l => l.OutputTokens),
+            TotalInputTokens = logs.Sum(l => (long)l.InputTokens),
+            TotalOutputTokens = logs.Sum(l => (long)l.OutputTokens),
             TotalCostUsd = logs.Sum(l => l.CostUsd),
             AvgResponseTimeMs = logs.Count > 0 ? logs.Average(l => l.ResponseTimeMs) : 0
         };
@@ -54,7 +54,7 @@ public class GetProjectUsageQueryHandler : IRequestHandler<GetProjectUsageQuery,
                 ModelType = g.Key.ModelType,
                 ModelTier = g.Key.ModelTier,
                 Requests = g.Count(),
-                Tokens = g.Sum(l => l.InputTokens + l.OutputTokens),
+                Tokens = g.Sum(l => (long)l.InputTokens + l.OutputTokens),
                 CostUsd = g.Sum(l => l.CostUsd)
             })
             .OrderByDescending(m => m.Requests)
